+++ src/main/scala/com/eneco/trading/kafka/connect/tennet/ImbalancePriceSourceRecordProducer.scala
@@ -13,7 +13,7 @@ case class ImbalancePriceSourceRecordProducer(readers: ServiceProvider, sourceTy
   override def mapRecord(record: Node, generatedAt: Long): Struct  = {
     new Struct(schema)
       .put("date", (record \ "DATE").text.toString)
-      .put("ptu", (record \ "PTU").text.toInt)
+      .put("ptu", (record \ "PTU").text.toLong)
       .put("period_from", (record \ "PERIOD_FROM").text.toString)
       .put("period_until", (record \ "PERIOD_UNTIL").text.toString)
       .put("upward_incident_reserve", TennetHelper.NodeSeqToDouble(record \ "UPWARD_INCIDENT_RESERVE"))
@@ -23,7 +23,7 @@ case class ImbalancePriceSourceRecordProducer(readers: ServiceProvider, sourceTy
       .put("incentive_component", TennetHelper.NodeSeqToDouble(record \ "INCENTIVE_COMPONENT"))
       .put("take_from_system", TennetHelper.NodeSeqToDouble(record \ "TAKE_FROM_SYSTEM"))
       .put("feed_into_system", TennetHelper.NodeSeqToDouble(record \ "FEED_INTO_SYSTEM"))
-      .put("regulation_state", (record \ "REGULATION_STATE").text.toInt)
+      .put("regulation_state", (record \ "REGULATION_STATE").text.toLong)
       .put("generated_at", generatedAt)
       .put("ptu_start", epochMillis.fromPTU((record \ "DATE").text.toString, (record \ "PTU").text.toInt))
   }
